Refactor app structure to support PWA scope isolation for external map view
- Moved main application routes to `app/app/` to allow creating a PWA scope at `/openroutes/app/`.
- Created `app/view.tsx` at the root (outside scope) to serve as the external "New Window" map view.
- Updated `app.json` `web.scope` and `web.startUrl` to `/openroutes/app/`.
- Updated `MapScreen.tsx` to open the external view URL instead of using the `standalone` parameter.
- Updated imports and layout files to reflect the new directory structure.
