+++ app.json
@@ -29,7 +29,8 @@
       "favicon": "./assets/favicon.png",
       "basePath": "openroutes",
       "display": "standalone",
-      "startUrl": "/openroutes/",
+      "scope": "/openroutes/app/",
+      "startUrl": "/openroutes/app/",
       "shortName": "OpenRoutes",
       "name": "OpenRoutes",
       "backgroundColor": "#FFFFFF",
