+++ app/app/(tabs)/setting.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SettingScreen from '../../components/screens/SettingScreen';
+import SettingScreen from '../../../components/screens/SettingScreen';
 
 export default function Layout(): React.ReactElement {
   return <SettingScreen />;
