+++ app/app/(tabs)/share.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ShareScreen from '../../components/screens/ShareScreen';
+import ShareScreen from '../../../components/screens/ShareScreen';
 
 export default function Layout(): React.ReactElement {
   return <ShareScreen />;
