+++ app/app/detail.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RouteDetailScreen from '../components/screens/RouteDetailScreen';
+import RouteDetailScreen from '../../components/screens/RouteDetailScreen';
 
 export default function Layout(): React.ReactElement {
   return <RouteDetailScreen />;
