+++ app/app/githubauth.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GithubAuthScreen from '../components/screens/GithubAuthScreen';
+import GithubAuthScreen from '../../components/screens/GithubAuthScreen';
 
 export default function Layout(): React.ReactElement {
   return <GithubAuthScreen />;
