+++ app/app/map.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useLocalSearchParams } from 'expo-router';
-import MapScreen from '../components/screens/MapScreen';
+import MapScreen from '../../components/screens/MapScreen';
 
 export default function MapRoute() {
   const { url, title, source, standalone } = useLocalSearchParams<{ url: string; title: string; source: string; standalone?: string }>();
