+++ app/index.tsx
@@ -1,0 +1,6 @@
+import { Redirect } from 'expo-router';
+import React from 'react';
+
+export default function Index() {
+  return <Redirect href="/app/" />;
+}
